bench: improve receive buffer usage
First, we were reading plaintext into an 8KB buffer.  That was bad,
because on the send side we left rustls to do the fragmentation
and naturally that means maximal-sized appdata messages (16KB).
So reading each such message involved:

- one 8KB copy into our buffer,
- a 8KB memmove by `alloc::vec::Drain::drop`
- a second 8KB copy into our buffer.

Having fixed that, we end up with multiple silly-sized stack buffers.
Using these for the first time tanks performance (because it will
fault, and the OS needs to get involved to grow our stack).  That caused
a very noticeable cost artifact in the first bulk test in any process.

Instead: put these buffers on the heap, in an object, and pass
them around in the tests rather than repeatedly allocating and
de-allocating them (at this size, most of the allocations are likely
to fall straight into mmap, which is slow).